+++ src/options/rule-editor/rule-visual-editor.ts
@@ -47,6 +47,7 @@ const DEFAULT_MATCHERS = [
   "context",
   "menuindex",
   "comment",
+  "directory",
   "linktext",
   "selectiontext",
   "referrerurl",